add command line API
