+++ bin/kysely-migration-cli.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('../dist/cli')
